# avantech.OrderingsSystem.Web/openapi/product-v1.yaml
openapi: 3.0.4
info:
  title: avantech.OrderingsSystem.Api
  version: '1.0'
paths:
  /:
    get:
      tags:
        - avantech.OrderingsSystem.Api
      responses:
        '200':
          description: OK
  /Product/products:
    get:
      tags:
        - Product
      operationId: GetProducts
      parameters:
        - name: productCategoryId
          in: query
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Product'
  /Product/insert:
    post:
      tags:
        - Product
      operationId: InsertProduct
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  '/Product/update/{id}':
    put:
      tags:
        - Product
      operationId: UpdateProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Product'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
  '/Product/delete/{id}':
    delete:
      tags:
        - Product
      operationId: DeleteProduct
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int32
      responses:
        '204':
          description: No Content
components:
  schemas:
    Product:
      type: object
      properties:
        productId:
          maxLength: 50
          minLength: 0
          type: integer
          format: int32
        name:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        price:
          type: number
          format: double
          nullable: true
        productCategoryId:
          type: integer
          format: int32
          nullable: true
      additionalProperties: false